Threads can be deleted only by their creators

+++ app/Http/Controllers/ThreadsController.php
@@ -111,6 +111,14 @@ class ThreadsController extends Controller
      */
     public function destroy(Channel $channel, Thread $thread)
     {
+        if ($thread->user_id != auth()->id()) {
+            if (request()->wantsJson()) {
+                return response(['status' => 'Permission denied.'], 403);
+            }
+
+            return redirect('/login');
+        }
+
         $thread->delete();
 
         if (request()->wantsJson()) {
